% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_meas.R
\name{CLPpred}
\alias{CLPpred}
\title{Category level prediction}
\usage{
CLPpred(CLPs = CLPs, Zs = Zs, p = p, n = n, lvls = lvls, datIN = datIN)
}
\arguments{
\item{CLPs}{Category level point coordinates}

\item{Zs}{Sample coordinates}

\item{p}{Number of variables}

\item{n}{Number of samples}

\item{lvls}{Names of category levels}

\item{datIN}{Input data from which \code{CLPs} and \code{Zs} are obtained}
}
\value{
\item{predCL}{Final predicted categorical data set}
}
\description{
Predicts category levels from an MCA based biplot using the distances between coordinates
}
