% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compdat}
\alias{compdat}
\title{Complete data example}
\format{
A data frame with 1000 rows and 5 columns.
}
\source{
Simulated data from a uniform distribution that is categorised into levels.
}
\description{
Simulated data example
}
\details{
\describe{
\item{V1}{Variable 1}
\item{V1}{Variable 2}
\item{V1}{Variable 3}
\item{V1}{Variable 4}
\item{V1}{Variable 5}
}
}
\keyword{datasets}
