% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPAbinFuncs.R
\name{delCL}
\alias{delCL}
\title{Deleting empty category levels}
\usage{
delCL(data)
}
\arguments{
\item{data}{is an object of the factor class}
}
\value{
\item{data}{Returns the input object, now with empty category levels removed}
}
\description{
This function removes unobserved category levels before applying multivariate techniques.
}
