% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DRT.R
\name{DRT}
\alias{DRT}
\title{Dimension reduction}
\usage{
DRT(missbp, method = c("MCA"))
}
\arguments{
\item{missbp}{An object of class \code{missbp} obtained from preceding function \code{missmi()}}

\item{method}{Select a dimension reduction technique. In the current version \code{MCA} is available.}
}
\value{
\item{Z}{List of sample coordinates}
\item{CLP}{List of category level point coordinates}
\item{lvls}{List of category level names}
\item{m}{Number of multiple imputations}
}
\description{
Multiple correspondence analysis is performed on the multiple imputed datasets
}
\examples{
data(implist)
missbp <- missmi(implist) |> DRT()

}
