% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggbiplFig}
\alias{ggbiplFig}
\title{ggBiplot function}
\usage{
ggbiplFig(
  missbp,
  Z.col = "#61223b",
  CLP.col = "#b79962",
  Z.pch = 19,
  CLP.pch = 15
)
}
\arguments{
\item{missbp}{An object of class \code{missbp} obtained from preceding function \code{missmi()}}

\item{Z.col}{Colour of sample coordinates}

\item{CLP.col}{Colour of category level point coordinates}

\item{Z.pch}{Plotting character of sample coordinates}

\item{CLP.pch}{Plotting character of category level point coordinates}
}
\value{
\itemize{
\item {plot}{Returns a GPAbin ggplot biplot.}
\item {plotC}{Returns an MCA ggplot biplot of the complete data set, if provided.}
}
}
\description{
Creates a multiple correspondence analysis (MCA) biplot in \code{ggplot}.
}
\examples{
data(implist)
data(compdat)
missbp <- missmi(implist)|> DRT() |> GPAbin() |> 
evalMeas(compdat = compdat) |> ggbiplFig()

### GPAbin biplot

missbp$plot

### MCA biplot

missbp$plotC

}
