% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{implist}
\alias{implist}
\title{List of multiple imputed data sets}
\format{
List containing five multiple imputations of \code{missdat}. Each list item a data frame with 1000 rows and 5 columns.
}
\source{
simulated example data imputed with mice::mice(missdat, m=5, method="polyreg", maxit=10, remove.collinear=FALSE, printFlag = FALSE)
}
\description{
Five multiple imputations of \code{missdat}
}
\details{
\describe{
\item{V1}{Variable 1}
\item{V1}{Variable 2}
\item{V1}{Variable 3}
\item{V1}{Variable 4}
\item{V1}{Variable 5}
}
}
\keyword{datasets}
