% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missmi.R
\name{missmi}
\alias{missmi}
\title{First step before constructing unified biplots}
\usage{
missmi(data)
}
\arguments{
\item{data}{input data frame or list}
}
\value{
\item{X}{The processed data}
\item{imputations}{Number of multiple imputations applied}
\item{n}{The number of samples}
\item{p}{The number of variables}
\item{miss_pct}{Percentage of missing values}
}
\description{
This function produces a list of elements to be used when producing a GPAbin biplot.
}
\examples{
data(missdat)
missbp <- missmi(missdat)
data(implist)
missbp <- missmi(implist)

}
