% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missmi.R
\name{print.missmi}
\alias{print.missmi}
\title{Generic print function for objects of class missmi}
\usage{
\method{print}{missmi}(x, ...)
}
\arguments{
\item{x}{an object of class \code{missmi}.}

\item{...}{additional arguments.}
}
\value{
This function will not produce a return value, it is called for side effects.
}
\description{
This function is used to print output when the missmi biplot object is created.
}
\examples{
data(missdat)
missbp <- missmi(missdat)
data(implist)
missbp <- missmi(implist)
print(missbp)

}
