% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Procrustes.R
\name{GPAbin}
\alias{GPAbin}
\title{Function to unify coordinates of multiple configurations}
\usage{
GPAbin(missbp, G.target = NULL)
}
\arguments{
\item{missbp}{An object of class \code{missmi} obtained from preceding function \code{missmi()}}

\item{G.target}{Target configuration. If not specified the centroid configuration will be used as the target.}
}
\value{
\item{Z.GPA.list}{List containing the sample coordinates for each MI after GPA}
\item{CLP.GPA.list}{List containing the CLPs for each MI after GPA}
\item{G.target}{Target configuration}
\item{Z.GPAbin}{Sample coordinates for the GPAbin biplot}
\item{CLP.GPAbin}{CLPs for the GPAbin biplot}
}
\description{
Combines multiple configurations from dimension reduction solutions applied to multiple imputed data sets
}
\examples{
data(implist)
missbp <- missmi(implist) |> DRT() |> GPAbin()

}
