% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{impute}
\alias{impute}
\title{Multiple imputation}
\usage{
impute(
  missbp,
  imp.method = c("MIMCA", "jomo", "DPMPM", "mice"),
  m = 5,
  dim = c("2D", "All")
)
}
\arguments{
\item{missbp}{An object of class \code{missmi} obtained from preceding function \code{missmi()}.}

\item{imp.method}{Select one of four imputation methods: \code{MIMCA}, \code{jomo}, \code{DPMPM}, \code{mice}}

\item{m}{Number of multiple imputations}

\item{dim}{Number of dimensions to use in final solutions (\verb{2D} or \code{All} available dimensions.)}
}
\value{
\item{dataimp}{List of imputed data}
}
\description{
Choose between four available multiple imputation strategies in \code{R}.
}
\examples{
\dontrun{
data(missdat)
missmi(missdat) |> impute(imp.method="DPMPM", m=5)}

}
